exec: execute and plan multi-col asc sorts
This commit adds the necessary infrastructure to support arbitrary
multi-column sorts. Only ascending sorts are possible right now, still.

The strategy is as follows. First, globally sort the first column,
keeping an ordinal column that is swapped as the first column is. Then,
for each subsequent column, perform the following steps:

1. find all partitions in the most recently sorted column, ORing them
   with the previous partition list if there is one
2. reorder the column to sort based on the ordinal order column
3. for each partition in the partition list, sort the column, updating
   the ordinal order column

This strategy seems to be pretty effective. The slowest thing about it
is still buffering the input data.

Benchmarks were updated so that nCols means # of sort columns instead #
of total columns.

```
BenchmarkSort/rows=2048/cols=1-8                  100000             19086 ns/op         858.39 MB/s
BenchmarkSort/rows=2048/cols=2-8                   30000             43096 ns/op         760.33 MB/s
BenchmarkSort/rows=2048/cols=4-8                   20000             82746 ns/op         792.01 MB/s
BenchmarkSort/rows=16384/cols=1-8                  10000            169373 ns/op         773.86 MB/s
BenchmarkSort/rows=16384/cols=2-8                   3000            407879 ns/op         642.70 MB/s
BenchmarkSort/rows=16384/cols=4-8                   2000            816472 ns/op         642.14 MB/s
BenchmarkSort/rows=262144/cols=1-8                   500           2729214 ns/op         768.41 MB/s
BenchmarkSort/rows=262144/cols=2-8                   200           6020773 ns/op         696.64 MB/s
BenchmarkSort/rows=262144/cols=4-8                   100          12384021 ns/op         677.37 MB/s
```

Release note: None

+++ pkg/sql/exec/coalescer.go
@@ -78,7 +78,7 @@ func (p *coalescerOp) Next() ColBatch {
 					bufferCol.CopyWithSelInt16(fromCol, sel[leftover:batchSize], batchSize-leftover, t)
 				} else {
 					toCol.Append(fromCol, t, uint64(p.group.Length()), leftover)
-					bufferCol.Copy(fromCol, int(leftover), int(batchSize), t)
+					bufferCol.Copy(fromCol, uint64(leftover), uint64(batchSize), t)
 				}
 			}
 		}
